address feedback, tidy up polyfills, build integration, docs
