+++ src/coreclr/nativeaot/BuildIntegration/Microsoft.NETCore.Native.targets
@@ -101,6 +101,7 @@ The .NET Foundation licenses this file to you under the MIT license.
     <SharedLibrary Condition="'$(_targetOS)' != 'win'">$(FrameworkLibPath)\libframework$(LibFileExt)</SharedLibrary>
     <IlcDynamicBuildPropertyDependencies Condition="'$(IlcCalledViaPackage)' == 'true'">SetupProperties</IlcDynamicBuildPropertyDependencies>
     <IlcDisableUnhandledExceptionExperience Condition="$(NativeCodeGen) == 'llvm'">true</IlcDisableUnhandledExceptionExperience>
+    <IlcLlvmTarget Condition="'$(NativeCodeGen)' == 'llvm' and '$(_targetOS)' == 'wasi'">wasm32-wasi-threads</IlcLlvmTarget>
   </PropertyGroup>
 
   <ItemGroup Condition="'$(NativeCodeGen)' == 'llvm'">
@@ -270,7 +271,7 @@ The .NET Foundation licenses this file to you under the MIT license.
       <IlcArg Condition="'$(TrimmerDefaultAction)' == 'copyused' or '$(TrimmerDefaultAction)' == 'copy' or '$(TrimMode)' == 'partial'" Include="--defaultrooting" />
       <IlcArg Include="--resilient" />
       <IlcArg Include="@(UnmanagedEntryPointsAssembly->'--generateunmanagedentrypoints:%(Identity)')" />
-      <IlcArg Condition="'$(_targetOS)' == 'wasi'" Include="--codegenopt:Target=wasm32-wasi-threads" />
+      <IlcArg Condition="'$(IlcLlvmTarget)' != ''" Include="--codegenopt:Target=$(IlcLlvmTarget)" />
 
       <IlcArg Condition="$(IlcDisableReflection) == 'true'" Include="--feature:System.Reflection.IsReflectionExecutionAvailable=false" />
 
@@ -329,8 +330,10 @@ The .NET Foundation licenses this file to you under the MIT license.
   </Target>
 
   <PropertyGroup Condition="'$(NativeCodeGen)' == 'llvm'">
-    <EmccOptimizationSetting Condition="$(Optimize) == 'true' and $(IlcOptimizationPreference) != 'Size'">-O3</EmccOptimizationSetting>
-    <EmccOptimizationSetting Condition="$(Optimize) == 'true' and $(IlcOptimizationPreference) == 'Size'">-Oz</EmccOptimizationSetting>
+    <WasmOptimizationSetting Condition="$(Optimize) == 'true' and $(IlcOptimizationPreference) != 'Size' and '$(_targetOS)' == 'browser'">-O3</WasmOptimizationSetting>
+    <!-- TODO-LLVM: test -O1/-O2/-O3 for size with Wasi SDK -->
+    <WasmOptimizationSetting Condition="$(Optimize) == 'true' and $(IlcOptimizationPreference) != 'Size' and '$(_targetOS)' == 'wasi'">-O2</WasmOptimizationSetting>
+    <WasmOptimizationSetting Condition="$(Optimize) == 'true' and $(IlcOptimizationPreference) == 'Size'">-Oz</WasmOptimizationSetting>
   </PropertyGroup>
 
   <Target Name="WasmObjects"
@@ -346,7 +349,7 @@ The .NET Foundation licenses this file to you under the MIT license.
       <CompileWasmArgs>$(CompileWasmArgs) -s DISABLE_EXCEPTION_CATCHING=0</CompileWasmArgs>
       <CompileWasmArgs Condition="'$(NativeDebugSymbols)' == 'true'">$(CompileWasmArgs) -g3</CompileWasmArgs>
       <!-- TODO-LLVM: remove -s USE_SDL=0 when we update to Emscripten with https://github.com/emscripten-core/emscripten/pull/18443 -->
-      <CompileWasmArgs>$(CompileWasmArgs) -s USE_SDL=0 $(EmccOptimizationSetting)</CompileWasmArgs>
+      <CompileWasmArgs>$(CompileWasmArgs) -s USE_SDL=0 $(WasmOptimizationSetting)</CompileWasmArgs>
 
       <ScriptExt Condition="'$(OS)' == 'Windows_NT'">.bat</ScriptExt>
       <WasmCompilerPath>&quot;$(EMSDK)/upstream/emscripten/emcc$(ScriptExt)&quot;</WasmCompilerPath>
@@ -355,7 +358,7 @@ The .NET Foundation licenses this file to you under the MIT license.
 
    <PropertyGroup Condition="'$(_targetOS)' == 'wasi'">
       <CompileWasmArgs>-fvisibility=default -mllvm -combiner-global-alias-analysis=false -mllvm -disable-lsr --sysroot=&quot;$(WASI_SDK_PATH)/share/wasi-sysroot&quot; -pthread -target wasm32-wasi-threads -c</CompileWasmArgs>
-      <CompileWasmArgs Condition="'$(NativeDebugSymbols)' != 'true'">$(CompileWasmArgs) -O2</CompileWasmArgs>
+      <CompileWasmArgs>$(CompileWasmArgs) $(WasmOptimizationSetting)</CompileWasmArgs>
       <CompileWasmArgs Condition="'$(NativeDebugSymbols)' == 'true'">$(CompileWasmArgs) -g3</CompileWasmArgs>
 
       <ExeExt Condition="'$(OS)' == 'Windows_NT'">.exe</ExeExt>
@@ -451,9 +454,10 @@ The .NET Foundation licenses this file to you under the MIT license.
       <CustomLinkerArg Include="@(NativeLibrary->'&quot;%(Identity)&quot;'->Replace(&quot;\&quot;, &quot;/&quot;))" />
       <CustomLinkerArg Include="@(NativeObjects->'&quot;%(Identity)&quot;'->Replace(&quot;\&quot;, &quot;/&quot;))" />
       <CustomLinkerArg Include="-o &quot;$(NativeBinary.Replace(&quot;\&quot;, &quot;/&quot;))&quot;" />
-      <CustomLinkerArg Condition="'$(NativeDebugSymbols)' != 'true'" Include="-O2 -flto" />
       <CustomLinkerArg Condition="'$(NativeDebugSymbols)' == 'true'" Include="-g3" />
+      <CustomLinkerArg Condition="'$(Optimize)' != 'true'" Include="$(WasmOptimizationSetting) -flto" />
       <CustomLinkerArg Condition="$(NativeLib) == 'Shared'" Include="-Wl,--export,NativeAOT_StaticInitialization" />
+      <CustomLinkerArg Condition="'$(IlcLlvmTarget)' != ''" Include="-target $(IlcLlvmTarget)" />
       <CustomLinkerArg Include="@(LinkerArg)" />
     </ItemGroup>
 
@@ -464,17 +468,12 @@ The .NET Foundation licenses this file to you under the MIT license.
       <CustomLinkerArg Include="-s ALLOW_MEMORY_GROWTH=1" />
       <CustomLinkerArg Include="-s DISABLE_EXCEPTION_CATCHING=0" />
       <CustomLinkerArg Include="-s ERROR_ON_UNDEFINED_SYMBOLS=0" />
-      <CustomLinkerArg Condition="'$(NativeDebugSymbols)' == 'true'" Include="-g3" />
-      <CustomLinkerArg Condition="'$(Optimize)' == 'true'" Include="-flto" />
-      <CustomLinkerArg Condition="'$(EmccOptimizationSetting)' != ''" Include="$(EmccOptimizationSetting)" />
-      <CustomLinkerArg Condition="$(NativeLib) == 'Shared'" Include="-Wl,--export,NativeAOT_StaticInitialization" />
-      <CustomLinkerArg Include="@(LinkerArg)" />
     </ItemGroup>
 
     <ItemGroup Condition  ="'$(_targetOS)' == 'wasi'" >
       <!-- Wasi has lots of undefined symbols currently, mostly globalization -->
       <CustomLinkerArg Include="-Wl,--unresolved-symbols=ignore-all -lstdc++ -Wl,--error-limit=0 -lpthread -pthread" />
-      <CustomLinkerArg Include="--sysroot=&quot;$(WASI_SDK_PATH.Replace(&quot;\&quot;, &quot;/&quot;))/share/wasi-sysroot&quot; -target wasm32-wasi-threads" />
+      <CustomLinkerArg Include="--sysroot=&quot;$(WASI_SDK_PATH.Replace(&quot;\&quot;, &quot;/&quot;))/share/wasi-sysroot&quot;" />
       <CustomLinkerArg Include="-Wl,--import-memory,--export-memory,--max-memory=2147483648" />
       <CustomLinkerArg Include="-lwasi-emulated-process-clocks -lwasi-emulated-signal -lwasi-emulated-mman -lwasi-emulated-getpid" />
     </ItemGroup>
@@ -485,7 +484,7 @@ The .NET Foundation licenses this file to you under the MIT license.
           Condition="('$(_targetOS)' == 'browser' and '$(EMSDK)' != '') or ('$(_targetOS)' == 'wasi' and '$(WASI_SDK_PATH)' != '')" />
     <Message Text="Emscripten not found, not linking WebAssembly. To enable WebAssembly linking, install Emscripten and ensure the EMSDK environment variable points to the directory containing upstream/emscripten/emcc.bat"
              Condition="'$(NativeCodeGen)' == 'llvm' and '$(_targetOS)' == 'browser'and '$(EMSDK)' == ''" Importance="High" />
-    <Message Text="Wasi SDK not found, not linking WebAssembly. To enable WebAssembly linking, install the Wasi SDK and ensure the WASI_SDK_PATH environment variable points to the directory containing bin/clang"
+    <Message Text="Wasi SDK not found, not linking WebAssembly. To enable WebAssembly linking, install the Wasi SDK and ensure the WASI_SDK_PATH environment variable points to the directory containing share/wasi-sysroot"
              Condition="'$(NativeCodeGen)' == 'llvm' and '$(_targetOS)' == 'wasi'and '$(WASI_SDK_PATH)' == ''" Importance="High" />
   </Target>
