+++ src/coreclr/nativeaot/Runtime/CMakeLists.txt
@@ -160,8 +160,8 @@ else()
     )
     if (CLR_CMAKE_TARGET_WASI)
       list(APPEND COMMON_RUNTIME_SOURCES
-        wasm/wasi.cpp
-        wasm/wasi.c
+        wasm/wasi_pthread.c
+        wasm/wasi_mmap.cpp
       )
     endif()
   endif()
